% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_estimates.R
\name{create_uk_hostorical_data}
\alias{create_uk_hostorical_data}
\title{Produce table of forecasts and actual data for all UK regions}
\usage{
create_uk_hostorical_data()
}
\description{
Fetch all the data from Azure Blob storage of Epiforecasts
and store in a directory in data.
}
