% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-wrang.R
\name{get_cases_forecasts_per_region}
\alias{get_cases_forecasts_per_region}
\title{Retrieve cases forcastes produced for all dates and all regions}
\usage{
get_cases_forecasts_per_region()
}
\description{
For all regions in a nations directory, extract the reported
cases and forcasted cases for regions and dates listed as subdirectories.
Files are saved to data/processed_data/uk_combined_actual_forecast.rds. Note
this assumes you have ran `create_uk_hostorical_data()`  and have the
data downloaded for processing.
}
