% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_estimates.R
\name{get_estimates}
\alias{get_estimates}
\title{Sync Estimates From Azure Storage
(taken from https://github.com/epiforecasts/covid-rt-estimates/blob/master/R/get_estimates.R)}
\usage{
get_estimates(
  local_dir,
  include = "csv",
  target_dir = "results",
  blob = "https://epinowcovidrstorage.blob.core.windows.net"
)
}
\arguments{
\item{local_dir}{A character string containing the local folder to sync
estimates into.}

\item{include}{A character vector indicating the results type to sync.
Options are
"csv", "rds", "png", and "log". The default is "csv" which represents a
summary.}

\item{target_dir}{A character string indicating the internal folder to
target in the  Azure blob storage. Defaults to "results". This may be used
to selectively sync results.}

\item{blob}{A character string indicating the Azure blob address.}
}
\description{
This utility allows for programmatic downloading/syncing of
results  from our reproduction number pipeline which are stored in Azure
blob storage.  It requires an installation of `azcopy` to function for
which we provide an install script for use with Linux. By default it will
download/sync all summary estimates. Optionally, it can also return plots
and samples (with short term archiving of previous estimates). Using the
file structure of `covid-rt-estimates` it can also be used to target
specific estimates by setting the `target_dir` argument.

`azcopy`: https://docs.microsoft.com/en-us/azure/storage/common/storage-use-azcopy-v10

Install file: bin/install_azcopy.sh
}
\examples{
get_estimates("covid-rt-estimates")
}
