% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-wrang.R
\name{get_forecasts_all_dates}
\alias{get_forecasts_all_dates}
\title{Retrieve forcastes produced for all dates for a given region}
\usage{
get_forecasts_all_dates(target_region)
}
\arguments{
\item{target_region}{char array path to a regions directory}
}
\description{
For a path to a given region's directory, extract the reported
cases and forcasted cases for all dates listed as subdirectories.
}
