% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-wrang.R
\name{get_forecasts_single_date}
\alias{get_forecasts_single_date}
\title{Retrieve forcastes produced for a given date at a given region}
\usage{
get_forecasts_single_date(target_region, target_date)
}
\arguments{
\item{target_region}{char array path to a regions directory}

\item{target_date}{char array date to extract data for.}
}
\description{
For a path to a given region's directory, extract the reported
cases and forcasted cases for a target date.
}
