% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-wrang.R
\name{get_rt_forecasts_single_date}
\alias{get_rt_forecasts_single_date}
\title{Retrieve RT forcastes produced for a given date at a given region}
\usage{
get_rt_forecasts_single_date(target_region, target_date, all = TRUE)
}
\arguments{
\item{target_region}{char array path to a regions directory}

\item{target_date}{char array date to extract data for.}

\item{all}{logical. Join all estimates or just the last one}
}
\description{
For a path to a given region's directory, extract the
forcasted RT for a target date.
}
